+++ RegisterPlayer.py
