+++ RegisterPlayerTests.py
@@ -1,4 +1,4 @@
-import register
+import RegisterPlayer
 import json
 import unittest
 
@@ -7,7 +7,7 @@ class TestStringMethods(unittest.TestCase):
 
     def test_registerPlayer(self):
         input = self.loadEvent("./test/resources/test-RegisterPlayer-input-01.json")
-        event = register.handler(input["input"], {})
+        event = RegisterPlayer.handler(input["input"], {})
         print(event)
         #self.assertEqual(event["data"]["game"]["gameState"], "running")
